+++ resources/views/centaur/travel_orders/index.blade.php
@@ -1,4 +1,4 @@
-<header class="page-header">
+<header class="page-header travel_header">
 	<div class="index_table_filter">
 		<label>
 			<input type="search"  placeholder="{{ __('basic.search')}}" onkeyup="mySearchTable()" id="mySearchTbl">
@@ -8,7 +8,20 @@
 				<i class="fas fa-plus"></i>
 			</a>
 		@endif
-		<span class="change_view"></span>
+		<select id="filter_employee" class="select_filter filter_travel" >
+			<option value="all">@lang('basic.all_employees')</option>
+			@foreach ($employees as $employee)
+				@if(count($employee->hasTravels) >0)
+					<option value="{{ $employee->first_name . ' ' .  $employee->last_name }}" class="id_{{ $employee->id }}">{{ $employee->first_name . ' ' .  $employee->last_name }}</option>
+				@endif
+			@endforeach
+		</select>
+		<select id="filter_date" class="select_filter filter_travel" >
+			<option value="all">@lang('basic.all_month')</option>
+			@foreach ($dates as $date)
+				<option value="{{ $date }}" class="date_{{ $date }}">{{ $date }}</option>
+			@endforeach
+		</select>
 	</div>
 </header>
 <main class="col-xs-12 col-sm-12 col-md-12 col-lg-12">
@@ -28,7 +41,7 @@
 				</thead>
 				<tbody>
 					@foreach ($travel_orders as $travel)
-						<tr>
+						<tr class="panel">
 							<td>{{ $travel->employee->user['first_name'] . ' ' .  $travel->employee->user['last_name'] }}</td>
 							<td>{{ $travel->date }}</td>
 							<td>{{ $travel->start_date }}</td>
@@ -36,17 +49,36 @@
 							<td>{{ $travel->destination }}</td>
 							<td>{{ $travel->car['registration'] }}</td>
 							<td class="center">
-								<!-- <button class="collapsible option_dots float_r"></button> -->
 								@if(Sentinel::getUser()->hasAccess(['travel_orders.update']))
-									<a href="{{ route('travel_orders.edit', $travel->id) }}" class="btn-edit" title="{{ __('basic.edit')}}" rel="modal:open">
-										<i class="far fa-edit"></i>
-									</a>
+									@if ($travel->status == 0 || $travel->status == null)
+										<a href="{{ route('travel_orders.edit', $travel->id) }}" class="btn-edit" title="{{ __('basic.edit')}}" rel="modal:open">
+											<i class="far fa-edit"></i>
+										</a>
+										<a href="{{ action('TravelOrderController@close_travel', ['id' => $travel->id]) }}" class="btn-finish close_travel" title="{{ __('basic.finish')}}" >
+											<i class="fas fa-check"></i>
+										</a>
+										<a href="{{ route('travelShow', $travel->id) }}" class="btn-edit" title="{{ __('basic.show')}}" target="_blank" >
+											<i class="far fa-eye"></i>
+										</a>
+									@else
+										<a href="{{ action('TravelOrderController@close_travel',  ['id' => $travel->id]) }}" class="btn-edit close_travel" title="{{ __('basic.open_order')}}" >
+											<i class="fas fa-times"></i>
+										</a>
+									@endif
 								@endif								
 								@if(Sentinel::getUser()->hasAccess(['travel_orders.delete']))
-									<a href="{{ route('travel_orders.destroy', $travel->id) }}" class="action_confirm btn-delete danger" title="{{ __('basic.delete')}}" data-method="delete" data-token="{{ csrf_token() }}">
-										<i class="far fa-trash-alt"></i>
-									</a>
+									@if ($travel->status == 0 || $travel->status == null)
+										<a href="{{ route('travel_orders.destroy', $travel->id) }}" class="action_confirm btn-delete danger" title="{{ __('basic.delete')}}" data-method="delete" data-token="{{ csrf_token() }}">
+											<i class="far fa-trash-alt"></i>
+										</a>
+									@endif
 								@endif
+								
+								@if ($travel->status == 1)
+									<a href="{{ asset('/travelOrder/Putni nalog_' . $travel->id . '.pdf') }}" target="_blank"><span class="pdf"></span></a>
+								
+								@endif
+								
 							</td>
 						</tr>
 					@endforeach
@@ -56,9 +88,13 @@
 			<p class="no_data">@lang('basic.no_data')</p>
 		@endif
 	</div>
+	<span hidden class="locale" >{{ App::getLocale() }}</span>
 </main>
 <script>
 	$(function(){
+		$.getScript( '/../js/filter_dropdown.js');
 		$.getScript( '/../js/filter_table.js');
+		$.getScript( '/../js/travel.js');
+		$.getScript( '/../restfulizer.js');
 	});
 </script>
